Supports defining namespace from IRB starup file
Based on development, there was duplication of
the name of the name of the pre-installed
instance variables.  They were both defined in
the binding namespace of the program as well in
the lib/latirb.rb file.

I have removed the bonds to the former, and thus
we now extract available names of instance
variables from lib/latirb.rb file.  Similarly,
when we need an instance of the variable, we pull
it out of the IRB::Context.workspace.binding.

In sum, this DRYs things out.