Refactor due to LatinVerb 0.9.5
Changes to LatinVerb (0.9.5) allow us to remove a
lot of strangeness that was used to accommodate
its metaprogrammatic design.

Principally, since we are no longer using
`method_missing` to derive valid TAB-complete-able
method names, we no longer need the gyrations that
required the creation of a custom completion
`Proc` (`Readline.completion_proc`).

Also, refactor the method inside of the `.begin`
method to enhance readability.